requirements.txt
fastapi
uvicorn
pikepdf
pdfplumber
reportlab
Pillow
pymupdf
python-docx
boto3